RUN: reorganise stuff
