+++ src/main/kotlin/org/rust/cargo/runconfig/RsRunner.kt
@@ -3,6 +3,7 @@ package org.rust.cargo.runconfig
 import com.intellij.execution.configurations.RunProfile
 import com.intellij.execution.executors.DefaultRunExecutor
 import com.intellij.execution.runners.DefaultProgramRunner
+import org.rust.cargo.runconfig.command.CargoCommandConfiguration
 
 class RsRunner : DefaultProgramRunner() {
     override fun canRun(executorId: String, profile: RunProfile): Boolean =
