+++ src/main/resources/META-INF/plugin.xml
@@ -421,14 +421,11 @@
         <!-- Run Configurations -->
 
         <programRunner implementation="org.rust.cargo.runconfig.RsRunner"/>
-        <configurationType implementation="org.rust.cargo.runconfig.CargoCommandRunConfigurationType"/>
-        <runConfigurationProducer
-            implementation="org.rust.cargo.runconfig.producers.CargoExecutableRunConfigurationProducer"/>
-        <runConfigurationProducer
-            implementation="org.rust.cargo.runconfig.producers.CargoTestRunConfigurationProducer"/>
+        <configurationType implementation="org.rust.cargo.runconfig.command.CargoCommandRunConfigurationType"/>
+        <runConfigurationProducer implementation="org.rust.cargo.runconfig.command.CargoExecutableRunConfigurationProducer"/>
+        <runConfigurationProducer implementation="org.rust.cargo.runconfig.test.CargoTestRunConfigurationProducer"/>
 
-        <runLineMarkerContributor language="Rust"
-                                  implementationClass="org.rust.cargo.runconfig.RsRunLineMarkerContributor"/>
+        <runLineMarkerContributor language="Rust" implementationClass="org.rust.cargo.runconfig.RsRunLineMarkerContributor"/>
 
         <!-- Cargo -->
