+++ src/test/kotlin/org/rust/cargo/runconfig/RunConfigurationTestCase.kt
@@ -9,6 +9,8 @@ import com.intellij.openapi.util.Disposer
 import com.intellij.openapi.util.Key
 import org.assertj.core.api.Assertions.assertThat
 import org.rust.cargo.RustWithToolchainTestBase
+import org.rust.cargo.runconfig.command.CargoCommandConfiguration
+import org.rust.cargo.runconfig.command.CargoCommandRunConfigurationType
 
 class RunConfigurationTestCase : RustWithToolchainTestBase() {
     override val dataPath = "src/test/resources/org/rust/cargo/runconfig/fixtures"
