+++ src/test/kotlin/org/rust/cargo/runconfig/filters/RsConsoleFilterTest.kt
@@ -1,4 +1,4 @@
-package org.rust.cargo.runconfig
+package org.rust.cargo.runconfig.filters
 
 class RsConsoleFilterTest : HighlightFilterTestBase() {
